delete done

+++ src/app/modules/header/header.controller.ts
@@ -150,7 +150,10 @@ const updateLogo = asyncHandler(async (req: Request, res: Response) => {
 const vanish = asyncHandler(async (req: Request, res: Response) => {
   const { id } = req.params;
   const header = await Header.findById({ _id: id });
-  header && (await deleteFromCloudinary(header.logo.src));
+  if (!header) {
+    throw new ApiError(404, "Header not found");
+  }
+  await deleteFromCloudinary(header.logo.src);
   const deletedHeader = await Header.deleteOne({ _id: id }).lean();
   resSender<IHeader>(res, {
     statusCode: 200,
